config: Refactor variable dereferencing
Instead of converting to a type right away, the variable is first fully
dereferenced (in the `dereference_any` method), and then converted to a
type in the `dereference` method. This enables us to implement
interpolation of variables, e. g. `height = ${vars.height}pt`.
Note that the `fallback` passed to `dereference()` is currently not
being used, this will be fixed later.