config: Introduce `get_literal` method
The method returns the literal value of a variable as a string.

Such a method is necessary, since the get<T> (which does a similar
thing) tries to dereference the value of the variable. For this reason,
`dereference_local` would dereference variables twice: Once in the the
call to `get<T>`, and then again in the call to `dereference`. So far
this behavior did not lead to any bugs (as far as I can tell), but if we
want to make variable reference escaping possible, it is necessary to
only dereference variable values once, since otherwise a string like
"$$$$" would be collapsed to just "$" (instead of "$$") because escaping
would be applied twice.