BROKEN: Escape variable references with backslash
This fails for the following case
```
[vars]
foo = foo
bar = \\${vars.foo}
```
The expected result is `\$foo`, but the actual result produced is
`$foo`. This happens because there's no way to know if a backslash has
been escaped at the stage of dereferencing, so the dereferencing method
thinks the `$` is escaped by the backslash. In order to make this work
it is necessary to dereference variables at the parsing stage, so that
we can differentiate escaped backslashes from non-escaped backslashes
preceding variable references (e. g. `\${vars.foo}\).